Refactor Android sync local store
